Move favorites api into favorites contrib app
