+++ geonode/api/api.py
@@ -18,7 +18,6 @@ from geonode.layers.models import Layer
 from geonode.maps.models import Map
 from geonode.documents.models import Document
 from geonode.groups.models import GroupProfile
-from geonode.contrib.favorite.models import Favorite
 
 from taggit.models import Tag
 from django.core.serializers.json import DjangoJSONEncoder
@@ -27,7 +26,6 @@ from tastypie import fields
 from tastypie.resources import ModelResource
 from tastypie.constants import ALL
 from tastypie.utils import trailing_slash
-from tastypie.authorization import Authorization
 
 from django.db.models import Q
 
@@ -120,37 +118,6 @@ class TagResource(TypeFilteredResource):
             'slug': ALL,
         }
         serializer = CountJSONSerializer()
-
-class FavoriteResource(ModelResource):
-    """Favorites API"""
-    type = fields.CharField()
-    detail_url = fields.CharField()
-    thumbnail_url = fields.CharField()
-
-    def dehydrate_type(self, bundle):
-      return bundle.obj.content_type
-
-    def dehydrate_detail_url(self, bundle):
-        if str(bundle.obj.content_type) == "layer":
-            return reverse('layer_detail', args=[Layer.objects.get(id=bundle.obj.id)])
-
-    def dehydrate_thumbnail_url(self, bundle):
-        if str(bundle.obj.content_type) == "layer":
-            return Layer.objects.get(id=bundle.obj.id).thumbnail_url
-
-    def serialize(self, request, data, format, options={}):
-        options['something'] = 'anything'
-        return super(FavoriteResource, self).serialize(request, data, format, options)
-
-    def authorized_read_list(self, object_list, bundle):
-        return object_list.filter(user=bundle.request.user.id)
-
-    class Meta:
-        queryset = Favorite.objects.all()
-        resource_name = 'favorites'
-        allowed_methods = ['get']
-        authorization = Authorization()
-
 
 class RegionResource(TypeFilteredResource):
     """Regions api"""
